pip does not compile the .c files yet
