+++ MANIFEST.in
@@ -1,0 +1,3 @@
+include lie_learn/representations/SO3/pinchon_hoggan/J_block_0-150.npy
+include lie_learn/representations/SO3/pinchon_hoggan/J_dense_0-150.npy
+global-include *.c
