+++ lie_learn/representations/SO3/pinchon_hoggan/download.py
@@ -7,5 +7,6 @@ def download(url):
     base = os.path.basename(url)
     path = os.path.join(os.path.dirname(__file__), base)
     if not os.path.isfile(path):
-        open(path, 'wb').write(requests.get(url).content)
+        with open(path, 'wb') as f:
+            f.write(requests.get(url).content)
     return np.load(path, encoding='latin1', allow_pickle=True)
