+++ lie_learn/representations/SO3/pinchon_hoggan/pinchon_hoggan_dense.py
@@ -1,10 +1,9 @@
+import os
+import numpy as np
+from scipy.linalg import block_diag
 
 # This code is not very optimized,
 # and can never become very efficient because it cannot exploit the sparsity of the J matrix.
-
-
-import numpy as np
-from scipy.linalg import block_diag
 
 # Load the J-matrices, which are stored in the same folder as this file
 from .download import download
@@ -14,7 +13,10 @@ from .download import download
 # Didier Pinchon1 and Philip E Hoggan2
 # https://iopscience.iop.org/article/10.1088/1751-8113/40/7/011/
 
-Jd = download('https://github.com/AMLab-Amsterdam/lie_learn/releases/download/v1.0/J_dense_0-278.npy')
+# Jd = download('https://github.com/AMLab-Amsterdam/lie_learn/releases/download/v1.0/J_dense_0-278.npy')
+base = 'J_dense_0-150.npy'
+path = os.path.join(os.path.dirname(__file__), base)
+Jd = np.load(path, allow_pickle=True)
 
 def SO3_irreps(g, irreps):
     global Jd
