--- lie_learn/spectral/tests/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'tsc'
