--- lie_learn/spectral/tests/test_conv_S2_SO3.py
@@ -1,8 +1,0 @@
-
-import lie_learn.spaces.S2 as S2
-from lie_learn.spaces.S3 import change_coordinates
-
-
-
-
-
