+++ setup.cfg
@@ -1,0 +1,7 @@
+[metadata]
+license_files = LICENSE
+
+[options]
+setup_requires =
+    numpy ; python_version>="3.0"
+    numpy<1.17 ; python_version<"3.0"
