+++ tests/spectral/test_conv_S2_SO3.py
@@ -1,0 +1,8 @@
+
+import lie_learn.spaces.S2 as S2
+from lie_learn.spaces.S3 import change_coordinates
+
+
+
+
+
